feat. 의미있는 함수끼리 그룹핑
